Move convert validation to util
