+++ lib/commands/convert.js
@@ -8,6 +8,7 @@ let shell = require('shelljs');
 let parser = require('gettext-parser');
 let AbstractCommand = require('./abstract');
 let BaseCommand = Object.create(AbstractCommand);
+let { validate } = require('./utils/validate-json');
 
 // enable error reporting
 shell.config.fatal = true;
@@ -119,7 +120,7 @@ CREATING JSON FILE
     let jsonData = parser.po.parse(poData);
     this._stripCommentsFromJSON(jsonData);
 
-    let validationErrors = this.validate(jsonData);
+    let validationErrors = validate(jsonData);
     this._printValidationErrors(validationErrors);
 
     if (this._shouldThrowForValidation(validationErrors, options.validateThrow)) {
@@ -151,177 +152,6 @@ CREATING JSON FILE
         let item = translations[namespace][k];
         delete item.comments;
       });
-    });
-  },
-
-  /**
-   * Validate a given JSON object for correct syntax.
-   * This returns an array of error objects.
-   *
-   * @method validate
-   * @param {Object} json
-   * @return {Object[]}
-   * @protected
-   */
-  _validationErrors: [],
-  validate(json) {
-    let { translations } = json;
-    this._validationErrors = [];
-
-    Object.keys(translations).forEach((namespace) => {
-      Object.keys(translations[namespace]).forEach((k) => {
-        this._validateItem(translations[namespace][k]);
-      });
-    });
-
-    return this._validationErrors;
-  },
-
-  /**
-   * Validate one translation item.
-   * This mutates this._validationErrors.
-   *
-   * @method _validateItem
-   * @param {Object }item
-   * @private
-   */
-  _validateItem(item) {
-    let { msgid: id, msgid_plural: idPlural, msgstr: translations } = item;
-
-    this._validatePlaceholders({ id, idPlural, translations });
-    this._validateTranslatedPlaceholders({ id, translations });
-  },
-
-  /**
-   * Validate the regular placeholders of an item.
-   * This possibly modifies this._validationErrors.
-   *
-   * @method _validatePlaceholder
-   * @param {String} id
-   * @param {String} idPlural
-   * @param {String[]} translations
-   * @private
-   */
-  _validatePlaceholders({ id, idPlural, translations }) {
-    // search for {{placeholderName}}
-    // Also search for e.g. Chinese symbols in the placeholderName
-    let pattern = /{{\s*(\S+?)\s*?}}/g;
-    let placeholders = id.match(pattern) || [];
-
-    // We also want to add placeholders from the plural string
-    if (idPlural) {
-      let pluralPlaceholders = idPlural.match(pattern) || [];
-      pluralPlaceholders.forEach((placeholder) => {
-        if (!placeholders.includes(placeholder)) {
-          placeholders.push(placeholder);
-        }
-      });
-    }
-
-    if (!placeholders.length) {
-      return;
-    }
-
-    translations.forEach((translation) => {
-      let translatedPlaceholders = translation.match(pattern) || [];
-
-      // Search for placeholders in the translated string that are not in the original string
-      let invalidPlaceholder = translatedPlaceholders.find((placeholder) => !placeholders.includes(placeholder));
-      if (invalidPlaceholder) {
-        this._validationErrors.push({
-          id,
-          translation,
-          message: `The placeholder "${invalidPlaceholder}" seems to be wrongly translated. Allowed: ${placeholders.join(', ')}`,
-          level: 'ERROR'
-        });
-      }
-    });
-  },
-
-  /**
-   * Validate the translated (complex) placeholders of an item.
-   * This mutates this._validationErrors.
-   *
-   * @method _validateTranslatedPlaceholders
-   * @param {String} id
-   * @param {String[]} translations
-   * @private
-   */
-  _validateTranslatedPlaceholders({ id, translations }) {
-    // find all: {{placeholder 'also translated'}}
-    // {{placeholder "also translated"}}
-    // {{placeholder \'also translated\'}}
-    // {{placeholder \"also translated\"}}
-    let pattern = /{{\s*([\w]+)\s+((\\')|(\\")|"|')(.*?)((\\')|(\\")|"|')\s*}}/g;
-
-    // This only works in non-plural form, so we assume only one translation
-    let [translation] = translations;
-
-    // Build an object describing the complex placeholders from the original string
-    let placeholders = id.match(pattern) || [];
-    if (!placeholders.length) {
-      return;
-    }
-    let placeholderConfig = placeholders.map((str) => {
-      pattern.lastIndex = 0;
-      let [fullResult, placeholder, quoteSymbol1, , , content, quoteSymbol2] = pattern.exec(str);
-      return {
-        fullResult,
-        placeholder,
-        quoteSymbol1,
-        quoteSymbol2,
-        content
-      };
-    });
-
-    // Build an object describing the complex placeholders from the translated string
-    let translatedPlaceholders = translation.match(pattern) || [];
-    let translatedPlaceholderConfig = translatedPlaceholders.map((str) => {
-      pattern.lastIndex = 0;
-      let [fullResult, placeholder, quoteSymbol1, , , content, quoteSymbol2] = pattern.exec(str);
-      return {
-        fullResult,
-        placeholder,
-        quoteSymbol1,
-        quoteSymbol2,
-        content
-      };
-    });
-
-    placeholderConfig.forEach(({ placeholder, content }) => {
-      // First we check for missing/invalid placeholders
-      // This can happen e.g. if a translator changes {{placeholder 'test'}} to {{placeholder `test`}}
-      // So we make sure that all originally defined placeholders actually still exist
-      if (!translatedPlaceholderConfig.find((config) => config.placeholder === placeholder)) {
-        this._validationErrors.push({
-          id,
-          translation,
-          message: `The complex placeholder "${placeholder}" is not correctly translated`,
-          level: 'ERROR'
-        });
-        return;
-      }
-
-      // Then, we check if the placeholder content is correctly translated
-      // If the whole string is not translated at all, we ignore it
-      // Only if the string is translated but the placeholder part not will this show a warning
-      // NOTE: This is just a warning (not an error), as it is theoretically possible this is done on purpose
-      // E.g. a word _might_ be the same in translated form
-      if (id === translation) {
-        return;
-      }
-      let invalidTranslatedPlaceholder = translatedPlaceholderConfig.find((config) => {
-        return config.content === content;
-      });
-
-      if (invalidTranslatedPlaceholder) {
-        this._validationErrors.push({
-          id,
-          translation,
-          message: `The content "${content}" for complex placeholder "${placeholder}" is not translated`,
-          level: 'WARNING'
-        });
-      }
     });
   },
